The beginning of a test class structure
tests can be written in the class and run with the testnumber (index of the array of function pointers) with a robotAxis and a degree.
In some cases (absolute move) will move to a home position and then test. In some cases (relative move) will move to a position based on the position it is currently in.